Adding docker file
