+++ src/main/java/com/projectgenerator/ai/genAiMicroservice/model/EntityModel.java
@@ -1,4 +1,4 @@
-package com.projectgenerator.ai.aiProjectGenertor.model;
+package com.projectgenerator.ai.genAiMicroservice.model;
 
 import java.util.List;
