+++ src/main/java/com/projectgenerator/ai/genAiMicroservice/service/CreateControllerService.java
@@ -1,7 +1,7 @@
-package com.projectgenerator.ai.aiProjectGenertor.service;
+package com.projectgenerator.ai.genAiMicroservice.service;
 
-import com.projectgenerator.ai.aiProjectGenertor.model.ProjectDetailsModel;
-import com.projectgenerator.ai.aiProjectGenertor.service.aiService.RepositoryPromptService;
+import com.projectgenerator.ai.genAiMicroservice.model.ProjectDetailsModel;
+import com.projectgenerator.ai.genAiMicroservice.service.aiService.ControllerPromptService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.stereotype.Service;
@@ -15,26 +15,30 @@ import java.util.Map;
 import java.util.Scanner;
 
 @Service
-public class CreateRepositoryService {
+public class CreateControllerService {
 
     @Value("${app.working-directory}")
     private String workingDirectory;
 
-    private RepositoryPromptService repositoryPromptService;
+    private final ControllerPromptService promptService;
+    private final CreateServiceClassService createServiceClassService;
 
-    public CreateRepositoryService(RepositoryPromptService repositoryPromptService){
-        this.repositoryPromptService=repositoryPromptService;
+    public CreateControllerService(ControllerPromptService promptService,
+                                   CreateServiceClassService createServiceClassService){
+        this.promptService=promptService;
+        this.createServiceClassService=createServiceClassService;
     }
 
-    public void createRepository(ProjectDetailsModel projectDetails,String serviceClassCode)throws IOException {
-        String code=repositoryPromptService.getRepositoryCode(projectDetails,serviceClassCode);
+    public void createControllerClass(ProjectDetailsModel projectDetails)throws IOException {
+        String code = promptService.generateControllerCode(projectDetails);
+
         Map<String, String> placeholders = Map.of(
-                "packageName", projectDetails.getGroupId(),
+                "packageName",projectDetails.getGroupId(),
                 "packageClass",projectDetails.getProjectName(),
-                "repository_code",code
-        );
-        System.out.println("Repository class response : "+code);
-        ClassPathResource resource = new ClassPathResource("templates/RepositoryTemplate.java");
+                "controller",code);
+        System.out.println("Controller class response : "+code);
+        createServiceClassService.createServiceClass(projectDetails,code);
+        ClassPathResource resource = new ClassPathResource("templates/ControllerTemplate.java");
         String content;
         try (InputStream inputStream = resource.getInputStream(); Scanner scanner = new Scanner(inputStream)) {
             content = scanner.useDelimiter("\\A").next(); // Read entire file as a string
@@ -53,16 +57,16 @@ public class CreateRepositoryService {
                 projectDetails.getProjectName(),
                 "src/main/java",
                 packagePath,
-                "repository");
+                "controller");
 
         Files.createDirectories(targetDir); // Ensure directory exists
 
         // Define target file path
-        Path targetPath = targetDir.resolve(projectDetails.getEntity().getEntityName() + "Repository.java");
+        Path targetPath = targetDir.resolve(projectDetails.getControllerModel().getControllerClassName() + ".java");
 
-        // Write processed content to the new repository class
+        // Write processed content to the new main class
         Files.writeString(targetPath, content, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING);
 
-        System.out.println("Repository Class created at: " + targetPath);
+        System.out.println("Controller Class created at: " + targetPath);
     }
 }
