+++ src/main/java/com/projectgenerator/ai/genAiMicroservice/service/CreateRepositoryService.java
@@ -1,6 +1,7 @@
-package com.projectgenerator.ai.aiProjectGenertor.service;
+package com.projectgenerator.ai.genAiMicroservice.service;
 
-import com.projectgenerator.ai.aiProjectGenertor.model.ProjectDetailsModel;
+import com.projectgenerator.ai.genAiMicroservice.model.ProjectDetailsModel;
+import com.projectgenerator.ai.genAiMicroservice.service.aiService.RepositoryPromptService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.stereotype.Service;
@@ -14,22 +15,26 @@ import java.util.Map;
 import java.util.Scanner;
 
 @Service
-public class CreateMainClassService {
+public class CreateRepositoryService {
 
     @Value("${app.working-directory}")
     private String workingDirectory;
 
-    public void createMainClass(ProjectDetailsModel projectDetails)throws IOException {
-        String className = projectDetails.getProjectName()
-                .substring(0,1)
-                .toUpperCase()+projectDetails.getProjectName()
-                .substring(1)+"Application";
+    private RepositoryPromptService repositoryPromptService;
+
+    public CreateRepositoryService(RepositoryPromptService repositoryPromptService){
+        this.repositoryPromptService=repositoryPromptService;
+    }
+
+    public void createRepository(ProjectDetailsModel projectDetails,String serviceClassCode)throws IOException {
+        String code=repositoryPromptService.getRepositoryCode(projectDetails,serviceClassCode);
         Map<String, String> placeholders = Map.of(
                 "packageName", projectDetails.getGroupId(),
                 "packageClass",projectDetails.getProjectName(),
-                "className", className);
-
-        ClassPathResource resource = new ClassPathResource("templates/MainClassTemplate.java");
+                "repository_code",code
+        );
+        System.out.println("Repository class response : "+code);
+        ClassPathResource resource = new ClassPathResource("templates/RepositoryTemplate.java");
         String content;
         try (InputStream inputStream = resource.getInputStream(); Scanner scanner = new Scanner(inputStream)) {
             content = scanner.useDelimiter("\\A").next(); // Read entire file as a string
@@ -47,16 +52,17 @@ public class CreateMainClassService {
         Path targetDir = Path.of(workingDirectory,
                 projectDetails.getProjectName(),
                 "src/main/java",
-                packagePath);
+                packagePath,
+                "repository");
 
         Files.createDirectories(targetDir); // Ensure directory exists
 
         // Define target file path
-        Path targetPath = targetDir.resolve(className + ".java");
+        Path targetPath = targetDir.resolve(projectDetails.getEntity().getEntityName() + "Repository.java");
 
-        // Write processed content to the new main class
+        // Write processed content to the new repository class
         Files.writeString(targetPath, content, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING);
 
-        System.out.println("Main Class created at: " + targetPath);
+        System.out.println("Repository Class created at: " + targetPath);
     }
 }
