+++ src/main/resources/application.properties
@@ -1,3 +1,4 @@
-spring.application.name=aiProjectGenertor
-app.working-directory=${HOME}/Downloads
-#app.working-directory=${WORKING_DIRECTORY:/app/downloads}
+spring.application.name=genAiMicroservice
+app.working-directory=${WORKING_DIRECTORY:/app/downloads}
+#app.working-directory=${HOME}/Downloads
+server.port=8182
