+++ src/test/java/com/projectgenerator/ai/genAiMicroservice/AiProjectGenertorApplicationTests.java
@@ -1,4 +1,4 @@
-package com.projectgenerator.ai.aiProjectGenertor;
+package com.projectgenerator.ai.genAiMicroservice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
